Fixed bug - Retrieving data with RowMapper<Integer> or <Double> was giving errors. Changed to the object of the same type as table and extracted price/distance from the returned object. Added calculate price method in 4 places so user can see price calculated after every relevant change made.

+++ project/src/main/java/com/example/project/Service/AddressService.java
@@ -42,7 +42,7 @@ public class AddressService {
     }
 
 
-    public int getDistanceFromId (int id) {
+    public double getDistanceFromId (int id) {
         return addressRepository.getDistanceFromId(id);
     }
 }
